refactor: move test files to feature folders
