+++ web/cypress/e2e/incidents/incidents_mocking_example.cy.ts
