+++ web/cypress/e2e/incidents/regression/end_and_resolution.cy.ts
