+++ web/cypress/e2e/incidents/regression/interval.cy.ts
