+++ web/cypress/e2e/incidents/regression/permission_and_silences.cy.ts
