+++ web/cypress/e2e/incidents/regression/stress_test_ui.cy.ts
