+++ web/cypress/e2e/incidents/regression/ui_interaction.cy.ts
