+++ web/cypress/e2e/incidents/regression/ui_regressions.cy.ts
