+++ web/cypress/e2e/legacy-dashboards/legacy_dashboards_regression.cy.ts
@@ -1,7 +1,7 @@
-import { runAllRegressionLegacyDashboardsTests } from '../../../support/monitoring/03.reg_legacy_dashboards.cy';
-import { runAllRegressionLegacyDashboardsTestsNamespace } from '../../../support/monitoring/06.reg_legacy_dashboards_namespace.cy';
-import { commonPages } from '../../../views/common';
-import { nav } from '../../../views/nav';
+import { runAllRegressionLegacyDashboardsTests } from '../../support/monitoring/03.reg_legacy_dashboards.cy';
+import { runAllRegressionLegacyDashboardsTestsNamespace } from '../../support/monitoring/06.reg_legacy_dashboards_namespace.cy';
+import { commonPages } from '../../views/common';
+import { nav } from '../../views/nav';
 
 const MP = {
   namespace: 'openshift-monitoring',
@@ -39,7 +39,7 @@ describe(
 // Test suite for Administrator perspective
 describe(
   'Regression: Monitoring - Legacy Dashboards Namespaced (Administrator)',
-  { tags: ['@legacy-dashboards'] },
+  { tags: ['@legacy-dashboards', '@metrics'] },
   () => {
     before(() => {
       cy.beforeBlock(MP);
