+++ web/cypress/e2e/metrics/metrics_virtualization_ivt.cy.ts
@@ -1,9 +1,11 @@
+import { runAllRegressionMetricsTests2 } from '../../support/monitoring/02.reg_metrics_2.cy';
 import { alerts } from '../../fixtures/monitoring/alert';
 import { runAllRegressionMetricsTests1 } from '../../support/monitoring/02.reg_metrics_1.cy';
 import { runAllRegressionMetricsTestsNamespace1 } from '../../support/monitoring/05.reg_metrics_namespace_1.cy';
 import { commonPages } from '../../views/common';
 import { nav } from '../../views/nav';
 import { guidedTour } from '../../views/tour';
+import { runAllRegressionMetricsTestsNamespace2 } from '../../support/monitoring/05.reg_metrics_namespace_2.cy';
 
 // Set constants for the operators that need to be installed for tests.
 const MCP = {
@@ -39,38 +41,16 @@ describe(
   () => {
     before(() => {
       cy.beforeBlockCOO(MCP, MP);
-    });
-
-    it('1. Installation: COO and setting up Monitoring Plugin', () => {
       cy.log('Installation: COO and setting up Monitoring Plugin');
-    });
-  },
-);
-
-describe(
-  'IVT: Monitoring UIPlugin + Virtualization',
-  { tags: ['@coo', '@slow', '@virtualization'] },
-  () => {
-    before(() => {
       cy.beforeBlockVirtualization(KBV);
-    });
-
-    it('1. Virtualization perspective - Observe Menu', () => {
       cy.log('Virtualization perspective - Observe Menu and verify all submenus');
-      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      cy.switchPerspective('Virtualization');
       guidedTour.closeKubevirtTour();
     });
-  },
-);
-
-describe(
-  'Regression: Monitoring - Metrics (Virtualization)',
-  { tags: ['@metrics', '@coo', '@slow', '@virtualization'] },
-  () => {
     beforeEach(() => {
       cy.visit('/');
       cy.validateLogin();
-      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      cy.switchPerspective('Virtualization');
       guidedTour.closeKubevirtTour();
       alerts.getWatchdogAlert();
       nav.sidenav.clickNavLink(['Observe', 'Metrics']);
@@ -92,7 +72,7 @@ describe(
     beforeEach(() => {
       cy.visit('/');
       cy.validateLogin();
-      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      cy.switchPerspective('Virtualization');
       guidedTour.closeKubevirtTour();
       alerts.getWatchdogAlert();
       nav.sidenav.clickNavLink(['Observe', 'Metrics']);
@@ -102,6 +82,58 @@ describe(
     });
 
     runAllRegressionMetricsTestsNamespace1({
+      name: 'Virtualization',
+    });
+  },
+);
+
+describe(
+  'Regression: Monitoring - Metrics (Virtualization)',
+  { tags: ['@metrics', '@slow', '@virtualization'] },
+  () => {
+    before(() => {
+      cy.beforeBlockCOO(MCP, MP);
+      cy.log('Installation: COO and setting up Monitoring Plugin');
+      cy.beforeBlockVirtualization(KBV);
+      cy.log('Virtualization perspective - Observe Menu and verify all submenus');
+      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      guidedTour.closeKubevirtTour();
+    });
+    beforeEach(() => {
+      cy.visit('/');
+      cy.validateLogin();
+      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      guidedTour.closeKubevirtTour();
+      alerts.getWatchdogAlert();
+      nav.sidenav.clickNavLink(['Observe', 'Metrics']);
+      commonPages.titleShouldHaveText('Metrics');
+      cy.changeNamespace('All Projects');
+      alerts.getWatchdogAlert();
+    });
+
+    runAllRegressionMetricsTests2({
+      name: 'Virtualization',
+    });
+  },
+);
+
+describe(
+  'Regression: Monitoring - Metrics Namespaced (Virtualization)',
+  { tags: ['@metrics', '@slow', '@virtualization'] },
+  () => {
+    beforeEach(() => {
+      cy.visit('/');
+      cy.validateLogin();
+      cy.switchPerspective('Virtualization', 'Fleet virtualization');
+      guidedTour.closeKubevirtTour();
+      alerts.getWatchdogAlert();
+      nav.sidenav.clickNavLink(['Observe', 'Metrics']);
+      commonPages.titleShouldHaveText('Metrics');
+      cy.changeNamespace(MP.namespace);
+      alerts.getWatchdogAlert();
+    });
+
+    runAllRegressionMetricsTestsNamespace2({
       name: 'Virtualization',
     });
   },
