+++ web/cypress/e2e/perses-dashboards/lightspeed_integration.cy.ts
