+++ web/cypress/e2e/perses-dashboards/perses_dashboards_edit.cy.ts
