+++ web/cypress/e2e/shared/admin_perspective_bvt.cy.ts
