+++ web/cypress/e2e/shared/coo_submenus.cy.ts
@@ -1,0 +1,57 @@
+import { commonPages } from '../../views/common';
+import { nav } from '../../views/nav';
+import { troubleshootingPanelPage } from '../../views/troubleshooting-panel';
+
+// Set constants for the operators that need to be installed for tests.
+const MCP = {
+  namespace: Cypress.env('COO_NAMESPACE'),
+  packageName: 'cluster-observability-operator',
+  operatorName: 'Cluster Observability Operator',
+  config: {
+    kind: 'UIPlugin',
+    name: 'monitoring',
+  },
+};
+
+const MP = {
+  namespace: 'openshift-monitoring',
+  operatorName: 'Cluster Monitoring Operator',
+};
+
+describe(
+  'BVT: COO',
+  {
+    tags: [
+      '@alerting',
+      '@cluster-health-analyzer',
+      '@perses-dashboards',
+      '@legacy-dashboards',
+      '@coo',
+    ],
+  },
+  () => {
+    before(() => {
+      cy.beforeBlockCOO(MCP, MP);
+    });
+
+    it('1. Admin perspective - Observe Menu', () => {
+      cy.log('Admin perspective - Observe Menu and verify all submenus');
+      cy.reload(true);
+      cy.wait(10000);
+      nav.sidenav.clickNavLink(['Observe', 'Alerting']);
+      commonPages.titleShouldHaveText('Alerting');
+      nav.tabs.switchTab('Silences');
+      nav.tabs.switchTab('Alerting rules');
+      nav.tabs.switchTab('Incidents');
+      nav.sidenav.clickNavLink(['Observe', 'Dashboards (Perses)']);
+      commonPages.titleShouldHaveText('Dashboards');
+      nav.sidenav.clickNavLink(['Observe', 'Alerting']);
+      troubleshootingPanelPage.openSignalCorrelation();
+      troubleshootingPanelPage.troubleshootingPanelPageShouldBeLoadedEnabled();
+    });
+
+    /**
+     * TODO: To be replaced by COO validation such as Dashboards (Perses) scenarios
+     */
+  },
+);
